backend/requirements.txt
flask>=3.0
flask-cors>=4.0
sqlalchemy>=2.0
psycopg2-binary>=2.9
marshmallow>=3.22
python-dotenv>=1.0
flasgger>=0.9.7
